+++ eosio-core/src/unsigned_int.rs
@@ -41,7 +41,7 @@ impl From<u8> for UnsignedInt {
 impl NumBytes for UnsignedInt {
     #[inline]
     fn num_bytes(&self) -> usize {
-        let mut val = self.0 as u64;
+        let mut val = u64::from(self.0);
         let mut bytes = 0_usize;
         loop {
             val >>= 7;
@@ -61,7 +61,7 @@ impl Read for UnsignedInt {
         let mut by = 0_u8;
         loop {
             let b = u8::read(bytes, pos)?;
-            v |= (((b & 0x7f) as u32) << by) as u64;
+            v |= u64::from(u32::from(b & 0x7f) << by);
             by += 7;
             if b & 0x80 == 0 {
                 break;
@@ -78,7 +78,7 @@ impl Write for UnsignedInt {
         bytes: &mut [u8],
         pos: &mut usize,
     ) -> Result<(), WriteError> {
-        let mut val = self.0 as u64;
+        let mut val = u64::from(self.0);
         loop {
             let mut b = (val as u8) & 0x7f;
             val >>= 7;
@@ -97,7 +97,7 @@ macro_rules! write_read_tests {
         #[cfg(test)]
         #[test]
         fn $i() {
-            let mut bytes = [0u8; 10];
+            let mut bytes = [0_u8; 10];
             let mut write_pos = 0;
             let varint: UnsignedInt = $v.into();
             assert_eq!(varint.num_bytes(), $n);
